Update `fallback-locale` example
