+++ examples/fallback-locale/src/routes/+layout.svelte
@@ -1,0 +1,25 @@
+<script>
+  import { writable } from 'svelte/store';
+  import { t, locale, locales } from '$lib/translations';
+
+  const count = writable(2);
+</script>
+
+<a href="/">{$t('menu.home')}</a>
+<a href="/about">{$t('menu.about')}</a>
+<br/>
+<br/>
+{$t('menu.notification', { count: $count })}<br />
+<button on:click="{() => {if ($count) $count -= 1;}}">–</button>
+<button on:click="{() => {$count += 1;}}">+</button>
+<hr />
+<slot />
+<br />
+<br />
+<br />
+<br />
+<select bind:value="{$locale}">
+  {#each $locales as value}
+    <option value="{value}">{$t(`lang.${value.toLowerCase()}`)}</option>
+  {/each}
+</select>
