+++ examples/fallback-locale/src/routes/+page.svelte
