+++ examples/fallback-locale/src/routes/about/+page.svelte
